Rename module state -> screen (#205)
